# backend/partenaire-ms/src/main/resources/partenaire-ms-openapi3.yaml
openapi: 3.0.0
info:
  title: Partenaire MS [WIP]
  description: Le microservice Partenaire MS permet de référencer tous les partenaires souhaitant intégrer le DNC, ainsi qu'à paraméter les configurations d'intégration DNC-Partenaire, choisir et configurer les démarches à réaliser ainsi que les données nécéssaires à leur réalisation   
  version: "1.0.0"
  
tags:
  - name: Authentification
    description: Mire d'authentification de l'utilisateur 
  - name: Partenaire 
    description: Opérations accessibles pour un partenaire habilité via IHM
  - name: Admin 
    description: Operations accessibles pour l'administrateur du DNC via IHM
  - name: Microservices DNC 
    description: Operations accessibles par les autres composants back-office 


paths:
  /account/integration:
    get:
      tags:
        - Admin
        - Partenaire
      summary: Récupération des paramètres d'intégration du partenaire
      description: Cet endpoint permet au partenaire de récupérer les paramètres d'intégration au DNC, à savoir partenaire_id,partenaire_secret, API_KEY, liste de couples <demarche_id><redirect_url_partenaire> 
      operationId: getIntegrationInfos
      security:
        - userJWT: []
      responses:
        '200': 
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationInfos'
        '400':
          description: Mauvaise requête 
        '401':
          description: Authorisation introuvable ou invalide
        '404':
          description: Ressource introuvable
        '498':
          description: Session expirée

    post:
      tags:
        - Admin
        - Partenaire
      summary: Modification des paramètres d'intégration d'une démarche existente
      description: Cet endpoint permet au partenaire de paraméter des infos d'intégration d'une démarche
      operationId: updateIntegrationInfos
      security:
        - userJWT: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrationInfos'
        description: Paramètres d'intégration modifiés
        required: true
      responses:
        '200': 
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationInfos'
        '400':
          description: Bad request
        '401':
          description: Authorization is missing or invalid
        
    put:
      tags:
        - Admin
        - Partenaire
      summary: Création d'une nouvelle démarche
      description: Cet endpoint permet au partenaire d'ajouter une démarche et de spécifier les paramètres d'intégration
      operationId: addIntegrationInfos
      security:
        - userJWT: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntegrationInfos'
        description: Paramètres d'intégration modifiés
        required: true
      responses:
        '200': 
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationInfos'
        '400':
          description: Bad request
        '401':
          description: Authorization is missing or invalid

  /account/integration/{id_demarche}:      
    delete:
      tags:
        - Admin
        - Partenaire
      summary: Suppression d'une démarche 
      description: Cet endpoint permet au partenaire de supprimer une démarche et de ses paramètres d'intégration
      operationId: deleteIntegrationInfos
      security:
        - userJWT: []
      parameters:
        - name: id_demarche
          in: path
          description: ID de la démarche
          required: true
          schema:
            type: string
      responses:
        '200': 
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationInfos'
        '400':
          description: Mauvaise requête 
        '401':
          description: Authorisation introuvable ou invalide
        '404':
          description: Ressource introuvable
        '498':
          description: Session expirée

  /partenaire:
    get:
      tags:
        - Microservices DNC
        - Admin
      summary: Liste de tous les partenaires existants
      description: Cet endpoint permet de récupérer la liste de tous les partenaires existants
      operationId: getListePartenaires
      security:
        - userJWT: []
      responses:
        '200': 
          description: Liste des partenaires disponibles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PartenaireInfos'
        '400':
          description: Mauvaise requête 
        '401':
          description: Authorisation introuvable ou invalide
        '404':
          description: Ressource introuvable
        '498':
          description: Session expirée

    put:
      tags:
        - Admin
      summary: Modifier un partenaire
      description: Cet endpoint permet à l'administrateur de modifier un partenaire
      operationId: updatePartenaire
      security:
        - userJWT: []
      parameters:
        - name: id_partenaire
          in: path
          description: ID du partenaire
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartenaireInfos'
        description: Information de mise à jour du partenaire
        required: true
      responses:
        '200': 
          description: Partenaire mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartenaireInfos'
        '400':
          description: Mauvaise requête 
        '401':
          description: Authorisation introuvable ou invalide
        '404':
          description: Ressource introuvable
        '498':
          description: Session expirée
        
    post:
      tags:
        - Admin
      summary: Créer un partenaire
      description: Cet endpoint permet à l'administrateur de créer un nouveau partenaire
      operationId: createPartenaire
      security:
        - userJWT: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartenaireInfos'
        description: Account added
        required: true
      responses:
        '200': 
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartenaireInfos'
        '400':
          description: Mauvaise requête 
        '401':
          description: Authorisation introuvable ou invalide
        '404':
          description: Ressource introuvable
        '498':
          description: Session expirée

        
  /partenaire/{siret_partenaire}:
    get:
      tags:
        - Microservices DNC
        - Admin
      summary: Accéder aux informations d'un partenaire
      description: Cet endpoint permet d'accéder aux information d'un partenaire existant
      operationId: getPartenaire
      security:
        - userJWT: []
      parameters:
        - name: siret_partenaire
          in: path
          description: SIRET du partenaire
          required: true
          schema:
            type: string
      responses:
        '200': 
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartenaireInfos'
        '400':
          description: Mauvaise requête 
        '401':
          description: Authorisation introuvable ou invalide
        '404':
          description: Ressource introuvable
        '498':
          description: Session expirée
    
    delete:
      tags:
        - Admin
      summary: Supprimer un compte partenaire
      description: Cet endpoint permet à l'administrateur de supprimer un compte d'un partenaire existant
      operationId: deletePartnerAccount
      security:
        - userJWT: []
      parameters:
        - name: id_partenaire
          in: path
          description: ID du partenaire
          required: true
          schema:
            type: string
      responses:
        '200': 
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartenaireInfos'
        '400':
          description: Mauvaise requête 
        '401':
          description: Authorisation introuvable ou invalide
        '404':
          description: Ressource introuvable
        '498':
          description: Session expirée

  /partenaire/{siret_partenaire}/credentials:
    put:
      tags:
        - Partenaire
      description: Création/Modification des credentials d'un partenaire
      operationId: saveCredentials
      parameters:
        - name: siret_partenaire
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        200:
          description: Message de bonne mise à jour du Partenaire et renvoie de l'identifiant des credentials.
          content:
            application/json:
              schema:
                type: string
        404:
          description: Partenaire inconnu
    post:
      tags:
        - Partenaire
      description: Création/Modification des credentials d'un partenaire
      operationId: updateCredentials
      parameters:
        - name: siret_partenaire
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
      responses:
        200:
          description: Message de bonne mise à jour du Partenaire
        401:
          description: L'ancien secret ne correspond pas au secret existant.
        404:
          description: Partenaire inconnu
  /partenaire/credentials/check:
    get:  
      tags: 
        - Partenaire
      summary: Vérifier les droits d'accès du partenaire
      description: Cet endpoint permet d'authentifier les accès d'un appel fait par le back-office partenaire pour accèder à une ressource, à travers le siret du partenaire et son paramètres client_id & client_secret
      operationId: checkCredentialsPartenaire
      security:
        - api_key: []
      parameters:
        - name: siret_partenaire
          in: query
          description: SIRET du partenaire
          required: true
          schema:
            type: string
        - name: client_id
          in: header
          description: ID du partenaire
          required: true
          schema:
            type: string
        - name: client_secret
          in: header
          description: ID du partenaire
          required: true
          schema:
            type: string
            format: password
      responses:
        '200': 
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GrantAccessPartner'
        '400':
          description: Mauvaise requête 
        '401':
          description: Authorisation introuvable ou invalide
        '404':
          description: Ressource introuvable
        '498':
          description: Session expirée
        


components:
  securitySchemes:
    api_key:
      type: apiKey
      name: X-API-Key
      in: header
  schemas:
    Credentials:
      description: Couple ClientID/ClientSecret à utiliser pour le partenaire
      type: object
      properties:
        clientId:
          description: Client ID des identifiants de connexion
          type: string
        clientSecret:
          description: Secret des identiants de connexion
          type: string
          format: password
        oldClientSecret:
          description: Ancien de l'identifiant de connexion, vide si c'est une création
          type: string
          format: password
    PartenaireInfos:
      type: object
      properties:
        siret:
          description: SIRET du partenaire
          type: string
        logo:
          type: string
          format: byte
          description: Logo rattaché à ce partenaire
        libelleCourt:
          type: string
          description: Libellé court rattachée à ce partenaire
        libelleLong:
          type: string
          description: Libellé long rattachée à ce partenaire
        adresse:
          type: string
        contact_responsable:
          type: object
          properties:
            nom_complet:
              type: string
            email: 
              type: string
        dateCreation:
          type: string
          description: Date de création du partenaire dans le DNC
          format: date-time
          example: '2016-08-29T09:12:33.001Z'
        dateMAJ:
          type: string
          description: Date de mise à jour du partenaire dans le DNC
          format: date-time
          example: '2016-08-29T09:12:33.001Z'

    IntegrationInfos:
      type: object
      required:
        - id
        - siret
        - urlTokenSend
      properties:
        id:
          type: string
          description: Identifiant du partenaire / Siret du partenaire
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        siret:
          type: string
          description: SIREN ou SIRET rattaché à ce partenaire
        urlTokenSend:
          type: string
          description: URL où le DNC envoie les tokens de démarches
        integration:
          type: array
          items:
            $ref: '#/components/schemas/IntegrationDemarche'
              
    IntegrationDemarche:
      type: object
      properties:
        demarche_id:
          type: string
          format: uuid
        demarche_libelle:
          type: string
        redirect_url:
          type: string
        demarche_meta:
          $ref: '#/components/schemas/DemarcheMetadata'
            
    DemarcheMetadata:
      type: object
      properties:
        scopes:
          type: array
          items:
            type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/DataSource'
          
    DataSource:
      type: object
      properties:
        fournisseur:
          type: string
        données:
          type: array
          items:
            type: string
        
    GrantAccessPartner:
      type: object
      properties:
        id:
          type: string